Fix regex capture in `sed` commands
Made sure that the proper amount of spaces (`\s`) are present when
searching for the text. Also made sure to match any extra characters
that come after ";", so it is possible to have comments afterwards.
Without this, any extra stuff outside the regex would be echoed.

Also added the `-r` flag in order to be able to make the regex a
little bit more readable.