Make certbot be the one to check if renewal is required
Previously we did a manual check of the certificates, to see if they
hadn't been modified for over a week and then we would trigger a
certbot renewal request. However, certbot would also check to see if
the certificates needed updating, before actually doing anything.

This manual check was done so we would not spam the Let's Encrypt
servers, and run into some limits, but this was before we noticed
that certbot doesn't seem to do any requests unless it has to.
Since its checks are much better, I see no need to have this double
checking stuff set up.