Add the `RENEWAL_INTERVAL` environmental variable
After we handed over the entire decision making, of whether to renew
a certificate or not, to certbot, it was now much easier to implement
an environmental variable which allows the user to decide how often
certbot should be triggered.

Certbot will still only renew certificates if they are within 30 days
of expiring, but with this addition we can let eager users trigger
certbot more often to make sure they always have an up to date
certificate.

The documentation and examples have also been updated to explain how
this new feture works.