Move certbot renewal loop code to its own file
I thought the `entrypoint.sh` was becoming a little bit too cluttered
so I took the liberty of moving all of the certbot related stuff
into the `run_certbot.sh` file. It also makes more sense to have all
that stuff in the same file for easier overview of what is happening.

There is no change in functionality by doing this move.