`wait` for both Nginx and the certbot update script
Previously we only cared about the Nginx process, and would kill
the container if that process died. The certbot update-loop did
very little before, so we basically ignored those rare instances
where it would actually be necessary to kill the container because
of fatal errors there. That meant we could land in a state where
Nginx still ran, but no certificates would be updated.

However, that is no longer the case, and the container will now
exit with the same status code as the first program to return
has. In practice this means that if any one of these processes
fail we will transfer its error code to the container's exit
code, and let the Docker deamon (+ admin) handle it from there.