Make it possible to run 'run_certbot.sh' as a standalone script
This change is made so that it should be possible to manually trigger
a reload of certbot, without restarting the container.

For this to be possible I had to move the "update loop" back to the
entrypoint script, but I think it makes more sense to have it there
after this change. Now it will be possible to run a "docker exec" on
the 'run_certbot.sh' script to manually trigger a renewal of the
certificates.