Reload configs on SIGHUP
When Nginx runs as PID 1, like in the parent Docker image, it listens
to the SIGHUP signal and reloads its configuration files without
having to restart the container. I wanted to make something similar
here, but we have the added responsibility of making sure all the
certificates and dhparam files are in place as well.

I think this is a pretty smooth way of handling reloading of the
configs, since with the killing of the sleep process we have sort
of a free "process lock" that makes sure we aren't making multiple
simultaneous request for the same certificate. Because if the
sleep process does not exist then we can be pretty sure that the
renewal loop is busy and has not reached the point where a new
sleep process has been spawned.

The annoying part of this solution is that apparently any "wait"
function in bash will return immediately if any "trap" is triggered,
which would mean that the container would exit with code 129 every
time we sent in SIGHUP. The ugly solution to this is that we just
loop around the final "wait", and check if the exit code that made
the "wait" return is 129. If it is we can be quite confident that
it is a signal created by a user, and it should be ignored.

Furthermore, we needed to catch the non-zero exit code of the
sleep function as well, so this code would not trip the "set -e"
setting we have in this child process. Since we kill it with a
"SIGTERM" we will just have the trailing "if" statement ignore
an exitcode of 143 (128+15), but issue an "exit" command for
anything else.

With all this in place we can finally handle a SIGHUP and
reload in the correct way.