Fix /entrypoint.d/ scripts not being triggered
The parent container's entrypoint.sh script had an extra safeguard
around itself, where it would not run any of the scripts inside the
/entrypoint.d/ folder, unless the container's CMD was either "nginx"
or "nginx-debug". This posed a problem for us, since we start it
with an entirely different command, which meant that none of the
startup scripts would trigger.

A solution to this is to just modify the `if` statement, inside the
entrypoint.sh script, so it triggers more easily. In this case we
change it so it will trigger on any command that contains "nginx"
anywhere in the string.