Create dhparams folder in case it is missing
In my examples I had forgotten that when you bind mount a folder,
then it will obscure the folder structure we have created inside
the container. This is not a problem for certbot, since it will
re-create its missing folders automatically, but our own makeshift
"dhparams" folder is not covered by it so we need to create it
manually.

I create this as a startup script, so people who do want to create
their own dhparams folders may just delete this file.