Add cargo package during build
Newer versions of the pip package "cryptography" make use of the rust
compiler when being built [1], so we will need the cargo package in
order for this to be successful.

This issue is primarily appearant in 32-bit ARM devices, which do not
have any pre-compiled wheels they can download, so it will have to
build it from source.

[1]: https://github.com/docker/compose/issues/8105#issuecomment-775055204