Make it possible to build for multiple architectures
Our parent container is build for multiple architectures, so I think
it would be nice if we followed suite. With this change we will be
able to build the following platforms:

- linux/amd64
- linux/386
- linux/arm64
- linux/arm/v7
- linux/mips64le
- linux/s390x
- linux/ppc64le

However, we will only create a push function for the top 4, since
the bottom 3 takes over 30 minutes to build, and none have asked
for these yet.

Instructions on how to set up Docker Buildx (which is required for
this) can be found in [issue #28][1].

The reason for the "32-bit fix" added to the Dockerfile is a very
interesting read, which can be found in [issue #30][2], but the
TL;DR is that there exist a very exotic bug for some low level
libraries when they try to access the filesystem while running inside
a 32-bit environment that is emulated by QEMU running on a 64-bit host.
For this reason we will need to pin the package version to something
that is not being compiled, but this is ONLY necessary if it is inside
a QEMU emulation.

[1]: https://github.com/JonasAlfredsson/docker-nginx-certbot/issues/28
[2]: https://github.com/JonasAlfredsson/docker-nginx-certbot/issues/30