Add GitHub workflow to build latest Docker image
This workflow will be triggered when something is changed inside either
the `src/` or the `.github/workflows/` folder on either the "main" or
the "master" branch, all other changes will be ignored. This should
be fine since an edit to the README would not affect how the Docker
container is being built.

More info about this can be found in issue #28:
https://github.com/JonasAlfredsson/docker-nginx-certbot/issues/28