Add GitHub workflow to build Docker image from tags
This workflow will be triggered when any tag, that match the semi-
regex (it is not normal regex, since "." is treated as a normal
period here), is pushed to the repository. The helper script will
then properly extract the version numbers we are interested in,
before the Docker image is built and tagged properly.

More info about this can be found in issue #28:
https://github.com/JonasAlfredsson/docker-nginx-certbot/issues/28