Add debug messages when `data:` or `engine:` are skipped
In a recent commit we made so that we do not check the validity of
a SSL filepath in case is started with `data:` or `engine:`. With
this change I want to add some descriptive log messages to what is
going on, in case someone tries to debug this in the future.