Create symlinker function from conf.d/ to user_conf.d/
In issue [#33][1] there was a request for having the container add
the configuration files at runtime, instead of having to build from
your own Dockerfile.

I was first sceptical of this, since this would deviate quite a bit
from the original setup of this image, but I think I have found a
method that is sort of backwards compatible while still allowing
people to just mount a local folder and have it work.

This will also allow us to update the config files currently in
nginx_conf.d, and have them propagate to users mounting folders
into the container.

[1]: https://github.com/JonasAlfredsson/docker-nginx-certbot/issues/33