Build more tags
With this change we introduce the ability to lock the version on
four different levels. The MAJOR tag will always point to the most
recent MINOR, which in turn points to the most recent PATCH, and
so on. This way users can determine how strict they want to be
when downloading updates for this image.