Fix potential race condition between Nginx start and symlinker
With the new symlinker service we might run into the situation that
it will manage to symlink over some files before Nginx starts its
process of loading in the configs.

This is not an optimal solution, and I will see if this can be solved
in a more robust fashion in the future.