Add Dockerfile that use the Alpine image
A user [wondered][1] why there was no Alpine version of this image,
so I took some time to see if this could be solved.

As of now it works, but I found a very interesting [bug][2] in Bash
(present in both Alpine and Debian) that might even be worth reporting
to those maintainers. However, for now we work around this one by using
an older version of Bash instead (<5.1.0).

[1]: https://github.com/JonasAlfredsson/docker-nginx-certbot/issues/35
[2]: https://github.com/JonasAlfredsson/bash_fail-to-wait