Only trigger pull request action on Dockerfile change
Since we currently do not have any tests for the bash code it is
pointless to trigger the GitHub Action that builds the Docker image,
since changes to anything else will have no impact on this. We
therefore limit to only trigger a build on pull request when they
introduce changes to the Dockerfile.