Drop i386 arch for Alpine
There are problems for the upstream images when they are trying to
build the i386 arch for Alpine, and there is no indication on when
this problem will be solved.

In order to not fall too far beind in Nginx updates we will drop
this arch for Alpine going forward.