Add support for a new cert:add_domains comment
This new comment allows for adding a domain in the certificates of the
current server block that wouldn't be declared directly in one of the
server_name entries. This allows for instance to use regular expressions
or suffix wildcards in the server_name block (which wouldn't be supported
by Let's Encrypt or any certificate authority) while indicating which
domains should actually be populated in the certificate.
e.g.
  server_name www.* homepage.*
  # cert:add_domains *.example.org *.example.com