Reduce complexity
The previous solution had a lot of features that increased the
complexity of the code significantly. I think that it might be
difficult to maintain it, and it made the barrier for entry for
newcomers higher. That is why we now remove some of the proposed
new features in favour of something that is a little bit easier
to follow.

We will lose the feature where the code automatically tried to
create wildcard certificates after all domain names it collected,
as well as it being a little bit more stupid since it cannot
differentiate between different server blocks in the same config
file. However, I think it is better if we force the user to
explicitly state (with the help of the '# certbot_domain' comment)
which domains they want to be included in the certificate request,
and I also think it is better for all if server blocks are split
up across more configuration files instead of us having to parse
and validate syntax.

What we are left with is now the ability to specify the same
cert_name in different configuration files and have the scripts
include all found server_names to this "key" in an associative
array instead. That way a user can reuse a wildcard certificate
over multiple configuration files by just using the same path in
all of them.

Furthermore, with the help of the '# certbot_domain' users now have
much more freedom in what they can put in the server_name field,
since the domain in the comment will take precedence. So we can now
have regex and other weird names without worrying that they will be
included in the request for certbot.