Remove workaround for cryptography package
The cryptography package fails to compile on 32-bit architectures
that are QEMU emulated on a 64-bit host. We had a workaround before
where we pinned the version to one that did compile without issues,
but some dependency now needs a newer version which braks everything
again.

Next type of workaround is now to change the filesystem to a tmpfs
one, which apparently does not trigger this bug. More information
is available in the issue:
https://github.com/JonasAlfredsson/docker-nginx-certbot/issues/30