Tune timeouts
It looks like both the Debian and Alpine builds take around 60 minutes
to complete (it is the arm/v7 arch that takes this long), so let's set
a timeout on 100 minutes. If it takes longer than that we can be quite
confident in that something is stuck and will probably need a restart.