Drop arm/v7 for Alpine builds again
This arch produces so many issues so I have no desire to keep this
maintained.

It is the building of the cryptography package that is that is the
main reason, but the error is often different:

 failed to fetch `https://github.com/rust-lang/crates.io-index`

 Caused by:
 - unexpected data at the end of the pack; class=Indexer (15)
 - failed to mmap. Could not write data: Out of memory; class=Os (2)
 - Value too large for defined data type; class=Os (2)

I have tried many workarounds, but don't really find it worth as
Alpine seems to move away from 32-bit either way. So we drop this
until it becomes stable enough that we don't have to do all of these
workarounds anymore.