Stop updating README on DockerHub
I use relative links in the README to refer to other files in this
repository, and those do not work when it is uploaded to Docker Hub.

To just make everything more simple for me I will just post a link
on Docker Hub pointing anyone interested in reading more to just
go to GitHub.

Resolves JonasAlfredsson/docker-nginx-certbot#180