Move certbot installation to requirements.txt
We have not really tracked which version of certbot that is installed
in the container, but I think we should. Adding it to the
requirements.txt file helps us get notified by dependabot when a new
version is available, and we can then trigger a rebuild to always keep
us up to date.

I also think we should try to bump at least our own patch version when
certbot moves. That way we allow users to also not have it change from
under their feet.