Add ability to change validity time of the local CA
A request to be able to more easily change this validity time was
raised when it came to deploying this container behind a CDN and
making it easy to set up an encrypted connection to the origin
server.

Through LOCAL_CA_ROOT_CERT_VALIDITY it should now be possible to
set it to any number of days desired, but we hide this information
inside the advanced documentation so we do falunt with it and
force users to read before blindly applying this in production.