Signal handler for SIGUSR1
This patch adds a signal handler for SIGUSR1 that forwards the signal to
nginx telling it to reopen log files. This is necessary after rotating
log files, for example. See https://nginx.org/en/docs/control.html#logs.

This patch also adds a section on signal handling under "Advanced Usage"
in the documentation.