Install latest rustc/cargo where needed
The cryptography Python package requires rustc >= 1.65.0 to build
now, and Debian only comes with 1.63.0 by default. We therefore
utilize the rustup installation script, similar to the get-pip.py
script, to get the latest version for the architectures where this
build step is needed.

In the Alpine case we never need to build the package, so remove
the rust dependency there.

In the case of i386 the rustup script is unable to correctly identify
which system it is to build for, so we need to provide this info
manually.